# openapi.yml
openapi: 3.0.1
info:
  title: mitum-currency - digest API
  description: >-
    This is API description of digest service of mitum-currency.
  contact:
    name: Spike^ekipS
    url: https://github.com/spikeekips/mitum-currency
    email: spikeekips@gmail.com
  license:
    name: GPL-3.0
    url: https://github.com/spikeekips/mitum-currency/blob/master/LICENSE
  version: 0.0.1
...
servers:
- url: https://localhost:54322
...
paths:
...
  /block/operations:
    get:
      tags:
      - block
      summary: All the operations
      operationId: operations
      parameters:
        - name: offset
          in: query
          schema:
            type: string
            example: "2,0"
          description: >-
            *operation*s after *offset*.
        - name: reverse
          in: query
          schema:
            type: boolean
            example: false
            default: false
          description: >-
            *operation*s by reverse order.
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "...."
                      detail:
                        type: string
                        example: "...."
        404:
          description: no more operations
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "operations not found"
                      detail:
                        type: string
                        example: "...."
        200:
          description: hal document of operations
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationsHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
  /account/{address}:
    get:
      tags:
      - account
      summary: The latest state of account
      description: >-
        The latest state of account. It contains the *keys* of account and it's *balance*.
      operationId: account
      parameters:
        - name: address
          in: path
          description: >
            *address* of account.
          required: true
          schema:
            $ref: '#/components/schemas/AccountAddress'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        404:
          description: not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of account
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/AccountHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
  /account/{address}/operations:
    get:
      tags:
      - account
      summary: Operations, which are related with the account
      description: >-
        Operations, which are related with the account.
      operationId: account-operations
      parameters:
        - name: address
          in: path
          description: >
            *address* of account.
          required: true
          schema:
            $ref: '#/components/schemas/AccountAddress'
        - name: offset
          in: query
          schema:
            type: string
            example: "2,0"
          description: >-
            *operation*s after *offset*.
        - name: reverse
          in: query
          schema:
            type: boolean
            example: false
            default: false
          description: >-
            *operation*s by reverse order.
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "...."
                      detail:
                        type: string
                        example: "...."
        404:
          description: no more operations
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "operations not found"
                      detail:
                        type: string
                        example: "...."
        200:
          description: hal document of operations
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/AccountOperationsHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
...
  /currency/{currency_id}:
    get:
      tags:
      - currency
      summary: Currency information
      operationId: currency
      parameters:
        - name: currency_id
          in: path
          description: currency unique id(or name)
          required: true
          schema:
            $ref: '#/components/schemas/CurrencyID'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of currency of *currency_id*
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CurrencyHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
components:
  schemas:
    Hint:
      type: string
      example: a015:0.0.1
    Problem:
      type: object
      required:
      - _hint
      - title
      - type
      - detail
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              example: a017:0.0.1
              default: a017:0.0.1
        title:
          type: string
          example: "sign"
        type:
          type: string
          example: "https://github.com/spikeekips/mitum-currency/problems/others"
        detail:
          type: string
          example: "invalid privatekey"
...
    HAL:
      type: object
      required:
      - _hint
      - _links
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              example: a016:0.0.1
              default: a016:0.0.1
        _embedded:
          type: object
          default: {}
        _links:
          type: object
          properties:
            self:
              allOf:
                - $ref: '#/components/schemas/HALLink'
                - type: object
                  properties:
                    href:
                      type: string
                      default: /
                      example: /
                    templated:
                      type: boolean
                      default: false
                      example: false
    HALLink:
      type: object
      required:
        - href
      properties:
        href:
          type: string
        templated:
          type: boolean
          description: The link is is templated link, not used without it's arguments.
          default: false
          example: false
...
    Height:
      description: >-
        height of *block*.
      type: integer
      format: int64
...
    AccountHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            hint:
              type: object
              properties:
                name:
                  type: string
                  default: mitum-currency-account-value
                  example: mitum-currency-account-value
                hint:
                  type: string
                  default: a018:0.0.1
                  example: a018:0.0.1
            _embedded:
              $ref: '#/components/schemas/AccountValue'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1
                operations:
                  description: >-
                    *operation*s, which are related of the account.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations
                operations:{offset}:
                  description: >-
                    *operation*s, which are related of the account after offset.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations?offset={offset}
                        templated:
                          type: boolean
                          default: true
                          example: true
                operations:{offset,reverse}:
                  description: >-
                    *operation*s, which are related of the account after offset by reverse order.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations?offset={offset&reverse=1}
                        templated:
                          type: boolean
                          default: true
                          example: true
                block:
                  description: >-
                    Request `/block/{height}`.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/244
...
    OperationHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            hint:
              type: object
              properties:
                name:
                  type: string
                  default: mitum-currency-operation-value
                  example: mitum-currency-operation-value
                hint:
                  type: string
                  default: a019:0.0.1
                  example: a019:0.0.1
            _embedded:
              $ref: '#/components/schemas/OperationValue'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /operation/6GymxmuvhgUfAKWYQUKKveYh1H8b981vQaZcMsbWqykS
                operation:{hash}:
                  description: request `/operation/{hash}`
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          default: /operation/{hash:(?i)[0-9a-z][0-9a-z]+}
                          example: /operation/{hash:(?i)[0-9a-z][0-9a-z]+}
                        templated:
                          type: boolean
                          default: true
                          example: true
                block:
                  description: request `/block/{height}` of the operation
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/244
                manifest:
                  description: request `/block/{height}/manifest` of the operation
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/244/manifest
    AccountOperationsHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            _embedded:
              type: array
              items:
                $ref: '#/components/schemas/OperationHAL'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations
                account:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1
                next:
                  description: >-
                    next operations with *offset*.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations?offset=2,0
                reverse:
                  description: >-
                    operations by reverse oder of self.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /account/B5ev8dDUpAdkCUnm8N2RQwUM86kcCLQqhCBd78FTxhtv-a000:0.0.1/operations?reverse=1
...
    OperationsHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            _embedded:
              type: array
              items:
                $ref: '#/components/schemas/OperationHAL'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations
                next:
                  description: >-
                    next operations with *offset*.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations?offset=2,0
                reverse:
                  description: >-
                    operations by reverse oder of self.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations?reverse=1
...
    CurrencyHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
             hint:
              type: object
              properties:
                name:
                  type: string
                  default: mitum-currency-currency-design
                  example: mitum-currency-currency-design
                hint:
                  type: string
                  default: a030:0.0.1
                  example: a030:0.0.1
             _embedded:
                $ref: '#/components/schemas/CurrencyDesign'
             _links:
                type: object
                properties:
                  self:
                    allOf:
                      - $ref: '#/components/schemas/HALLink'
                      - type: object
                        properties:
                          href:
                            type: string
                            default: /currency/XXX
                            example: /currency/XXX
    CurrencyID:
      description: currency unique id(or name)
      type: string
      example: XXX
    AccountValue:
      allOf:
        - $ref: '#/components/schemas/Account'
        - type: object
          required:
          - balance
          - height
          - previous_height
          properties:
            balance:
              $ref: '#/components/schemas/Amount'
            height:
              $ref: '#/components/schemas/Height'
            previous_height:
              $ref: '#/components/schemas/Height'
...
    AccountAddress:
      description: account address
      type: string
      example: 'mc-node-010a:0.0.1'
...